Added a function to check for updated levels.
The levels are now stored online at http://pyspy.game-host.org/levels.
They are checked via an md5 hash. This function simply checks if the md5
sums of the .xcf files on the server are different to the local ones.

Next task is to download levels if they have been updated.