spyThis is now fully playable with this commit.
The level selector needs work, so although it's playable it might not be
that fun. The list of masks under the image are now removed one by one as
they are found. A tick is drawn to show that a click has been successful,
new level transitions implemented and many bug fixes from the last commit.